+++ src/main/java/org/hrds/rdupm/harbor/api/vo/HarborQuotaVo.java
@@ -25,7 +25,7 @@ public class HarborQuotaVo {
 	private Integer usedCount;
 
 	@ApiModelProperty("存储容量限制值")
-	private Integer storageLimit;
+	private Long storageLimit;
 
 	@ApiModelProperty(value = "存储容量限制--数值")
 	private Integer storageNum;
@@ -34,7 +34,7 @@ public class HarborQuotaVo {
 	private String storageUnit;
 
 	@ApiModelProperty("存储容量已使用值")
-	private Integer usedStorage;
+	private Long usedStorage;
 
 	@ApiModelProperty(value = "已使用存储容量--数值")
 	private BigDecimal usedStorageNum;
